# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Assemble Yourself",
        "description": "Приложение для личностного роста",
        "contact": {
            "email": "v.a.kuznetsov-83@yandex.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "profile"
                ],
                "summary": "Информация о пользователе",
                "description": "Получение информации о пользователе по токену",
                "operationId": "getProfileByToken",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "readOnly": "true",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "Имя пользователя",
                                            "type": "string",
                                            "readOnly": "true"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "readOnly": "true",
                                            "example": "user1@mail.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "format": "password",
                                            "minLength": 8,
                                            "readOnly": "true",
                                            "example": "PassWord12345"
                                        },
                                        "about": {
                                            "type": "string",
                                            "readOnly": "true",
                                            "example": "Обо мне"
                                        },
                                        "Courses": {
                                            "$ref": "#/components/schemas/CourseModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "profile"
                ],
                "summary": "Изменение профиля пользователя",
                "description": "Изменение профиля пользователя. Пользователь должен быть авторизован",
                "operationId": "updateProfile",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UserArray"
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Авторизуйтесь для изменения профиля"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Переданы не все обязательные поля пользователь с таким email уже существует",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Поле id обязательно для заполнения"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CourseModel": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "course_name": {
                        "description": "Название курса",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "description": {
                        "description": "Описание курса",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "tag": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "cover_url": {
                        "description": "URL адрес обложки",
                        "type": "string"
                    },
                    "author": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "start_date": {
                        "description": "Дата начала курса",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "end_date": {
                        "description": "Дата окончания курса",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "course_program": {
                        "description": "Содержание программы курса",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "deleted_at": {
                        "description": "Soft delete timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Courses"
                }
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user1@mail.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "readOnly": "true",
                        "example": "PassWord12345"
                    },
                    "about": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Обо мне"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Courses"
                }
            }
        },
        "requestBodies": {
...
            "UserArray": {
                "description": "Перечень свойств пользователя",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": "1"
                                },
                                "name": {
                                    "description": "Имя пользователя",
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "user1@mail.com"
                                },
                                "password": {
                                    "type": "string",
                                    "format": "password",
                                    "minLength": 8,
                                    "example": "PassWord12345"
                                },
                                "about": {
                                    "type": "string",
                                    "example": "Обо мне"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer {Token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}